Powermenu compatibility update for OpenRC & SystemD

+++ .config/rofi/power_menu.sh
@@ -13,6 +13,12 @@ options=" Shutdown\0icon\x1fsystem-shutdown
  Logout\0icon\x1fsystem-log-out
  Cancel\0icon\x1fprocess-stop"
 
+if command -v systemctl >/dev/null 2>&1; then
+    INIT=systemd
+else
+    INIT=openrc
+fi
+
 # Show the menu and get user choice
 chosen=$(echo -e "$options" | rofi -dmenu -i -p "System Actions" \
     -show-icons \
@@ -24,27 +30,46 @@ chosen=$(echo -e "$options" | rofi -dmenu -i -p "System Actions" \
 # Execute the chosen action
 case $chosen in
     " Shutdown")
-        systemctl poweroff
+        if [[ $INIT == systemd ]]; then
+            systemctl poweroff
+        else
+            sudo halt
+        fi
         ;;
     " Reboot")
-        systemctl reboot
+        if [[ $INIT == systemd ]]; then
+            systemctl reboot
+        else
+            sudo reboot
+        fi
         ;;
     " Lock Screen")
-        # Change this to your preferred screen locker
-        # Common options: i3lock, xscreensaver-command -lock, gnome-screensaver-command -l
+        # Cambiar al locker que uses
         i3lock-fancy
         ;;
     " Suspend")
-        systemctl suspend
+        if [[ $INIT == systemd ]]; then
+            systemctl suspend
+        else
+            sudo pm-suspend
+        fi
         ;;
     " Hibernate")
-        systemctl hibernate
+        if [[ $INIT == systemd ]]; then
+            systemctl hibernate
+        else
+            sudo pm-hibernate
+        fi
         ;;
     " Logout")
-        # Simple and elegant logout - try universal methods
-        loginctl terminate-session "${XDG_SESSION_ID:-self}" 2>/dev/null || \
-        loginctl kill-session "${XDG_SESSION_ID:-self}" 2>/dev/null || \
-        pkill -TERM -s -1
+        if [[ $INIT == systemd ]]; then
+            loginctl terminate-session "${XDG_SESSION_ID:-self}" 2>/dev/null || \
+            loginctl kill-session "${XDG_SESSION_ID:-self}" 2>/dev/null || \
+            pkill -TERM -s -1
+        else
+            # OpenRC / sesión X11 simple
+            pkill -KILL -u "$USER"
+        fi
         ;;
     " Cancel")
         exit 0
